Add support for setting breakpoints in a library by file address
If you're deep in asm land and want to set a breakpoint in a file you're debugging, it's tedious to set a breakpoint as you have to do `image list`, find the address of the library. Then open up Calculator.app and do the math. (And sometimes the maths are off because of confusing load addresses!)

This lets you do the easy thing by just specifying an offset as you'd see in your disassembler.

Test plan:
```
(lldb) binside 0
Breakpoint 2: where = LaunchServices`LaunchServices[0x0000000000000000], address = 0x00007fff88c60000
```