Provide a help string when registering commands
I don't know when it was added, but `command script add` has a --help
flag, which defines the help text to show in the output of the `help`
command.

This change calls the description method on each command, extracts the
first line, and then passes that as the help string when registering the
command.