Add findinstances, and new support framework in Chisel.xcodeproj
This adds a new command, `findinstances`, which is implemented mostly in native
code and so comes with Chisel.xcodeproj which contains a new Chisel framework.
Future Chisel commands can implement the heavy lifing this way too.

The `findinstances` command scans the heap using available iOS/macOS malloc API.
For each allocation, a number of heuristics are peformed to identify likely
Objective-C instances. The heuristics do not call methods on the objects,
instead relying only on objc runtime functions to passively match the instance
based on its class metadata. This avoids allocations and stateful side effects
in the objc runtime.

After this first pass, the candidate objects go through a second pass that
checks if they match against an optional `NSPredicate`. If there's no predicate,
the object is printed with minimal information. If there is a predicate, and the
object passes the predicate, then the object will be printed out with more
detail, specifically the detail queried in the predicate.