Taplog: print the touch's UIGestureRecognizers
Summary:
I think more information here will make everyone's lives easier while debugging!

It's common for gesture recognizers to preempt or otherwise interfere with a hit-testing view's touch handling:
- A UIScrollView can delay touches began (and does so with a private GR)
- A Tap GR on the target view or its superview can handle a touch and prevent touchesBegain from ever getting called on the hit-testing view (e.g. before the `delaysTouchesBegain` delay expires)
- A hit-testing UIButton (and only a UIButton) can cause a tap gesture recognizer to fail

etc.

Maybe we can also follow up with a mechanism for breaking when a given GR's state changes. But in the meantime it'd at least be helpful to add a watch expression and inspect a given GR manually.

Reviewed By: samishchandra

Differential Revision: D20908394

fbshipit-source-id: 89d075cc4073573013ef5fe41ebcaa026a77e2e5