Copy command
Summary:
The `copy` command can be used to copy any `NSURL` or `NSData` to the host machine.

The item will be copied to `/tmp/chisel_copy` and opened in the default application. If copying an `NSURL`, the file will retain its existing name and extension. If copying an `NSData`, a name will be generated with the `.data` extension.

Options:
- `--filename`: Set a name for the file to be copied.
- `--no-open`: Prevent the file from being opened after copying.

Reviewed By: kolinkrewinkel

Differential Revision: D20922370

fbshipit-source-id: a4491f2ef03bf115fce37abdb79e2bd7269635a8