docs: add GH button in support of Ukraine (#302)
Summary:
Our mission at [Meta Open Source](https://opensource.facebook.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

Pull Request resolved: https://github.com/facebook/chisel/pull/302

Reviewed By: kolinkrewinkel

Differential Revision: D34648105

Pulled By: dmitryvinn-fb

fbshipit-source-id: 382dd09f1372c175d3861631ca3503d89566450d