Send command output to lldb result object
Summary:
- Print Chisel command output even if user isn't seeing STDOUT

## Details:

According to the [lldb documentation](https://fburl.com/znu9pep9), the `result` object in a custom lldb command is:

> A return object which encapsulates success/failure information for the command and output text **that needs to be printed as a result of the command**. The plain Python “print” command also works but text won’t go in the result by default (it is useful as a temporary logging facility).

Prior to this change, we were sending any Chisel command output to STDOUT via `print()`, `parse_args`, etc., and NOT writing to the `result` object in Chisel commands. This works fine if STDOUT is visible to the user, but in some deployments of lldb, STDOUT is not used (or is sent to dev/null), which breaks the Chisel output.

With this change, we make sure we're always writing output to the result object so Chisel doesn't break in environments where STDOUT is not visible to the user.

Reviewed By: kolinkrewinkel

Differential Revision: D35554193

fbshipit-source-id: 908f1beeca78bded8b55cac219bfed593c46ae17